# src/main/java/com/system/SmallBusinessBookingSystem/controller/dto/AuthCredentialsDto.java
package com.system.SmallBusinessBookingSystem.controller.dto;

import com.fasterxml.jackson.databind.PropertyNamingStrategies;
import com.fasterxml.jackson.databind.annotation.JsonNaming;
import lombok.Builder;
import lombok.Data;

@Data
@Builder
@JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
public class AuthCredentialsDto {
    private final String email;
    private final String password;
}