# src/main/java/com/system/SmallBusinessBookingSystem/controller/dto/BookingCreateDto.java
package com.system.SmallBusinessBookingSystem.controller.dto;

import com.fasterxml.jackson.databind.PropertyNamingStrategies;
import com.fasterxml.jackson.databind.annotation.JsonNaming;
import lombok.Builder;
import lombok.Data;

import java.time.Instant;

@Data
@Builder
@JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
public class BookingCreateDto {
    private final String userId;
    private final String serviceId;
    private final Instant dateTime;
}