# src/main/java/com/system/SmallBusinessBookingSystem/controller/dto/BusinessCreateDto.java
package com.system.SmallBusinessBookingSystem.controller.dto;

import com.fasterxml.jackson.databind.PropertyNamingStrategies;
import com.fasterxml.jackson.databind.annotation.JsonNaming;
import lombok.Builder;
import lombok.Data;

@Data
@Builder
@JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
public class BusinessCreateDto {
    private final String businessName;
    private final String description;
}