# src/main/java/com/system/SmallBusinessBookingSystem/controller/dto/BusinessUpdateDto.java
package com.system.SmallBusinessBookingSystem.controller.dto;

import com.fasterxml.jackson.databind.PropertyNamingStrategies;
import com.fasterxml.jackson.databind.annotation.JsonNaming;
import lombok.Builder;
import lombok.Data;

@Data
@Builder
@JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
public class BusinessUpdateDto {
    private final String id;
    private final String businessName;
    private final String description;
}