# src/main/java/com/system/SmallBusinessBookingSystem/controller/dto/NotificationCreateDto.java
package com.system.SmallBusinessBookingSystem.controller.dto;

import com.fasterxml.jackson.databind.PropertyNamingStrategies;
import com.fasterxml.jackson.databind.annotation.JsonNaming;
import com.system.SmallBusinessBookingSystem.service.models.NotificationType;
import lombok.Builder;
import lombok.Data;

import java.time.Instant;

@Data
@Builder
@JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
public class NotificationCreateDto {
    private final String message;
    private final Instant sentAt;
    private final NotificationType type;
}