# src/main/java/com/system/SmallBusinessBookingSystem/controller/dto/PasswordResetCompleteDto.java
package com.system.SmallBusinessBookingSystem.controller.dto;

import com.fasterxml.jackson.databind.PropertyNamingStrategies;
import com.fasterxml.jackson.databind.annotation.JsonNaming;
import lombok.Builder;
import lombok.Data;

@Data
@Builder
@JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
public class PasswordResetCompleteDto {
    private final String token;
    private final String password;
}